Renamed function / Added connection string logic / Instantiated client outside of the function
